+++ src/main/java/ru/vyarus/guice/persist/orient/db/user/SpecificUserAction.java
@@ -2,10 +2,14 @@ package ru.vyarus.guice.persist.orient.db.user;
 
 /**
  * Action performed with specific user (different from default one).
- * Must be passed into
- * {@link ru.vyarus.guice.persist.orient.db.user.UserManager#executeWithUser(String, String, SpecificUserAction)}.
- * <p>Specific user must be set before transaction opening. By implementing this interface and calling
- * user change method will not start any implicit transaction.</p>
+ * <p>When used to set user outside of transaction
+ * {@link ru.vyarus.guice.persist.orient.db.user.UserManager#executeWithUser(String, String, SpecificUserAction)},
+ * affects connection user for all transactions created inside callback (used when multiply connections
+ * must be affected).</p>
+ * <p>When used to set user inside transaction
+ * {@link ru.vyarus.guice.persist.orient.db.user.UserManager#executeWithTxUser(String, SpecificUserAction)},
+ * affects only current connection user. This allows to use user specific logic like security ro current user
+ * assigning to record.</p>
  *
  * @param <T> return type (may be Void)
  * @author Vyacheslav Rusakov
@@ -14,7 +18,6 @@ package ru.vyarus.guice.persist.orient.db.user;
 public interface SpecificUserAction<T> {
 
     /**
-     *
      * @return return value (or null)
      * @throws Throwable any error thrown during execution
      */